AristaEAPI.jl Documentation

AristaEAPI.jl Documentation

High level interface

Low level interface

Index

High level interface

The high level interface consists of functions and structures that make it easy and convenient to query the switch for statistics and status information and work with the returned data.

Functions

The high level interface provides query functions and predicate functions. Query functions communicate with the switch to get statistics and status information. Predicate functions can be used with functions that take a predicate function, such as filter, findall, etc.

Query functions

These function return the results as a Vector of structures. Each functions has a corresponding structure defined for it. The returned Vectors are compatible with Tables.jl, so they can be used with any Table.jl consumer, such as PrettyTables.jl When used with PrettyTables, pretty_table methods are defined to alter some default keyword argument values to display all entries and provide more tailored alignment (as shown in the example below).

AristaEAPI.interfaces_counters_rates — Function

interfaces_counters_rates(host[, interfaces]; username, password, protocol)

Return Vector{InterfacesCountersRates} of interface counter rates from host.

interfaces may be given as a comma separated string of interfaces or a Vector of interfaces to be queried. If not given, all interfaces will be queried.

username and password default to admin. protocol defaults to https, but may be given as http instead.

AristaEAPI.interfaces_transceiver — Function

interfaces_transceiver(host[, interfaces]; username, password, protocol)

Return NamedTuple of interface transceiver status from host.

interfaces may be given as a comma separated string of interfaces or a Vector of interfaces to be queried. If not given, all interfaces will be queried.

username and password default to admin. protocol defaults to https, but may be given as http instead.

AristaEAPI.lldp_neighbors — Function

lldp_neighbors(host[, interfaces]; username, password, protocol)

Return Vector{LLDPNeighbors} of LLDP neighbors of host.

interfaces may be given as a comma separated string of interfaces or a Vector of interfaces to be queried. If not given, all interfaces will be queried.

username and password default to admin. protocol defaults to https, but may be given as http instead.

AristaEAPI.mac_address_table — Function

mac_address_table(host[, interfaces]; address, username, password, protocol)

Return Vector{MACAddressRecord} of the unitcast MAC address table of host.

interfaces may be given as a comma separated string of interfaces or a Vector of interfaces to be queried. If not given, all interfaces will be queried. Optionally, a MAC address may be passed as address to query for a specific MAC address.

username and password default to admin. protocol defaults to https, but may be given as http instead.

Example

julia> interfaces_counters_rates("myswitch", "et4/1-et5/4")|>pretty_table
┌─────────────┬─────────────┬──────────┬────────────────┬──────────────┬──────────────┬─────────┬─────────────┬──────────┐
│ port        │ description │ interval │         in_bps │      in_util │       in_pps │ out_bps │    out_util │  out_pps │
├─────────────┼─────────────┼──────────┼────────────────┼──────────────┼──────────────┼─────────┼─────────────┼──────────┤
│ Ethernet4/1 │             │      300 │    9.10519e9   │ 91.4426      │  2.44134e5   │ 21967.1 │ 0.000232733 │  8.16335 │
│ Ethernet4/2 │             │      300 │    9.1052e9    │ 91.4426      │  2.44134e5   │ 21947.4 │ 0.000232518 │  8.1528  │
│ Ethernet4/3 │             │      300 │    9.10519e9   │ 91.4426      │  2.44134e5   │ 21944.4 │ 0.000232479 │  8.14663 │
│ Ethernet4/4 │             │      300 │    9.1052e9    │ 91.4426      │  2.44134e5   │ 21944.6 │ 0.000232483 │  8.1481  │
│ Ethernet5/1 │             │      300 │    5.03642e-17 │  5.31133e-25 │  1.71821e-20 │ 21686.8 │ 0.000229122 │  7.65873 │
│ Ethernet5/2 │             │      300 │    9.1052e9    │ 91.4426      │  2.44134e5   │ 21955.6 │ 0.000232627 │  8.16933 │
│ Ethernet5/3 │             │      300 │ 5682.48        │  7.307e-5    │ 10.1532      │ 32662.7 │ 0.000370916 │ 27.6806  │
│ Ethernet5/4 │             │      300 │    9.1052e9    │ 91.4426      │  2.44134e5   │ 21943.1 │ 0.000232465 │  8.14637 │
└─────────────┴─────────────┴──────────┴────────────────┴──────────────┴──────────────┴─────────┴─────────────┴──────────┘

Predicate functions

These predicate functions are intended for use with filter, findall, etc.

AristaEAPI.isdynamic — Function

isdynamic(mac::MACAddressRecord) -> Bool

Predicate function that returns mac.type == "dynamic

AristaEAPI.isstatic — Function

isstatic(mac::MACAddressRecord) -> Bool

Predicate function that returns mac.type == "static

Structures

The structures defined by AristaEAPI.jl are documented here.

AristaEAPI.InterfacesCountersRates — Type

struct InterfacesCountersRates <: AbstractAristaEAPIStruct
    port::String
    description::Union{String, Missing}
    interval::Union{Int, Missing}
    in_bps::Union{Float64, Missing}
    in_util::Union{Float64, Missing}
    in_pps::Union{Float64, Missing}
    out_bps::Union{Float64, Missing}
    out_util::Union{Float64, Missing}
    out_pps::Union{Float64, Missing}
end

Structure of interface rates returned by interfaces_counters_rates.

Fields

Name	Description

port	Interface name

description	Interface description

interval	Averagine interval

in_bps	Inbound bits per second

in_util	Inbound utilization percentage

in_pps	Inbound packets per second

out_bps	Outbound bits per second

out_util	Outbound utilization percentage

out_pps	Outbound packets per second

AristaEAPI.InterfacesTransceiver — Type

struct InterfacesTransceiver <: AbstractAristaEAPIStruct
    port::String
    temperature::Union{Float64, Missing}
    voltage::Union{Float64, Missing}
    bias_current::Union{Float64, Missing}
    tx_power::Union{Float64, Missing}
    rx_power::Union{Float64, Missing}
    last_update::Union{Float64, Missing}
    media_type::Union{String, Missing}
    vendor_sn::Union{String, Missing}
    narrowband::Union{Bool, Missing}
end

Structure of transceiver status info returned by interfaces_transceiver.

Fields

Name	Description

port	Interface name

temperature	Transceiver temperature (C)

voltage	Transceiver voltage (V)

bias_current	Transceiver bias current (mA)

tx_power	Optical transmit power (dBm)

rx_power	Optical receive power (dBm)

last_update	Last update time (UNiX seconds)

media_type	Media type (e.g. 40GBASE-SR4)

vendor_sn	Transceiver serial number

narrowband	Narrowband transceiver

AristaEAPI.LLDPNeighbors — Type

struct LLDPNeighbors <: AbstractAristaEAPIStruct
    port::Union{String, Missing}
    neighbor_device::Union{String, Missing}
    neighbor_port::Union{String, Missing}
    ttl::Union{Int, Missing}
end

Structure of LLDP neighbor info returned by lldp_neighbors.

Fields

Name	Description

port	Interface name

neighbor_device	Neighbor name

neighbor_port	Neighbor interface

ttl	Time to live

AristaEAPI.MACAddressRecord — Type

struct MACAddressRecord <: AbstractAristaEAPIStruct
    vlan::Union{Int, Missing}
    mac_address::Union{String, Missing}
    type::Union{String, Missing}
    port::Union{String, Missing}
    moves::Union{Int, Missing}
    last_move::Union{Float64, Missing}
end

Structure of MAC address info returned by mac_address_table.

Fields

Name	Description

vlan	VLAN ID

mac_address	MAC address

type	Indicates learned ("dynamic") or configured ("static")

port	Interface name

moves	Number of times the MAC address has changed ports

last_move	Last time the MAC address changed ports (UNIX seconds)

Low level interface

These functions are used by the high level functions to run commands on the switch using the Arista EAPI. They can be used to run any commands that your switch supports over EAPI. This function returns a JSON3.Object that can be used as a Dict{Symbol, Any}. This object represents the JSON RPC response from the switch. The results of interest can be obtained from this object via the :result key.

AristaEAPI.run_command — Function

run_command(host, cmd::AbstractString; username, password, protocol)

Run Arista command cmd on host using Arista EAPI.

username and password default to admin. protocol defaults to https, but may be given as http instead.

AristaEAPI.run_commands — Function

run_commands(host, cmds::AbstractVector; username, password, protocol)

Run Arista commands cmds on host using Arista EAPI.

username and password default to admin. protocol defaults to https, but may be given as http instead.

Example

julia> reply = run_command("arista1g", "show version")
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 3 entries:
  :jsonrpc => "2.0"
  :result  => Object[{…
  :id      => "AristaEAPI.jl-yA6T96mC"

julia> reply[:result]
1-element JSON3.Array{JSON3.Object, Base.CodeUnits{UInt8, String}, SubArray{UInt64, 1, Vector{UInt64}, Tuple{UnitRange{Int64}}, true}}:
 {
          "modelName": "DCS-7048T-A-R",
    "internalVersion": "4.15.7M-3284043.4157M",
   "systemMacAddress": "00:1c:73:a0:b8:df",
       "serialNumber": "JPE14471431",
           "memTotal": 3978148,
    "bootupTimestamp": 1.75581673972e9,
            "memFree": 770812,
            "version": "4.15.7M",
       "architecture": "i386",
      "isIntlVersion": false,
    "internalBuildId": "b0b0dff8-c9ca-40cc-a625-7fd3c8c76ebd",
   "hardwareRevision": "01.07"
}

Index

AristaEAPI.InterfacesCountersRates

AristaEAPI.InterfacesTransceiver

AristaEAPI.LLDPNeighbors

AristaEAPI.MACAddressRecord

AristaEAPI.interfaces_counters_rates

AristaEAPI.interfaces_transceiver

AristaEAPI.isdynamic

AristaEAPI.isstatic

AristaEAPI.lldp_neighbors

AristaEAPI.mac_address_table

AristaEAPI.run_command

AristaEAPI.run_commands