AristaEAPI.jl Documentation

AristaEAPI.jl Documentation

High level functions

Low level function

Index

High level functions

These functions return the results of commonly used commands for getting switch status and statistics. All of them return a Vector{NamedTuple}. The keys of the NamedTuples closely follow the field names used in the textual output of the switch CLI.

AristaEAPI.interfaces_counters_rates — Function

interfaces_counters_rates(host[, interfaces]; username, password, protocol)

Return NamedTuple of interface counter rates from host.

interfaces may be given as a comma separated string of interfaces or a Vector of interfaces to be queried. If not given, all interfaces will be queried.

username and password default to admin. protocol defaults to https, but may be given as http instead.

AristaEAPI.lldp_neighbors — Function

lldp_neighbors(host[, interfaces]; username, password, protocol)

Return NamedTuple of LLDP neighbors of host.

interfaces may be given as a comma separated string of interfaces or a Vector of interfaces to be queried. If not given, all interfaces will be queried.

username and password default to admin. protocol defaults to https, but may be given as http instead.

AristaEAPI.mac_address_table — Function

mac_address_table(host[, interfaces]; address, username, password, protocol)

Return NamedTuple of the unitcase mac address table of host.

interfaces may be given as a comma separated string of interfaces or a Vector of interfaces to be queried. If not given, all interfaces will be queried. Optionally, a MAC address may be passed as address to query for a specific MAC address.

username and password default to admin. protocol defaults to https, but may be given as http instead.

Low level function

Thes functions are used by the high level functions to run commands on the switch using the Arista EAPI. They can be used to run any commands that your switch supports over EAPI. This function returns a JSON3.Object that can be used as a Dict{Symbol, Any}. This object represents the JSON RPC response from the switch. The results of interest can be obtained from this object via the :result key.

AristaEAPI.run_command — Function

run_command(host, cmd::AbstractString; username, password, protocol)

Run Arista command cmd on host using Arista EAPI.

username and password default to admin. protocol defaults to https, but may be given as http instead.

AristaEAPI.run_commands — Function

run_commands(host, cmds::AbstractVector; username, password, protocol)

Run Arista commands cmds on host using Arista EAPI.

username and password default to admin. protocol defaults to https, but may be given as http instead.

Example

julia> reply = run_command("arista1g", "show version")
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 3 entries:
  :jsonrpc => "2.0"
  :result  => Object[{…
  :id      => "AristaEAPI.jl-yA6T96mC"

julia> reply[:result]
1-element JSON3.Array{JSON3.Object, Base.CodeUnits{UInt8, String}, SubArray{UInt64, 1, Vector{UInt64}, Tuple{UnitRange{Int64}}, true}}:
 {
          "modelName": "DCS-7048T-A-R",
    "internalVersion": "4.15.7M-3284043.4157M",
   "systemMacAddress": "00:1c:73:a0:b8:df",
       "serialNumber": "JPE14471431",
           "memTotal": 3978148,
    "bootupTimestamp": 1.75581673972e9,
            "memFree": 770812,
            "version": "4.15.7M",
       "architecture": "i386",
      "isIntlVersion": false,
    "internalBuildId": "b0b0dff8-c9ca-40cc-a625-7fd3c8c76ebd",
   "hardwareRevision": "01.07"
}

Index

AristaEAPI.interfaces_counters_rates

AristaEAPI.lldp_neighbors

AristaEAPI.mac_address_table

AristaEAPI.run_command

AristaEAPI.run_commands